+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.9.1 (2025-08-10)
 
 - Added more entries to PATH in the ARexx execution environment for so it ARexx can run more commands just from their name
+- Improve error handling in chat responses
 
 ## 2.9.0 (2025-08-09)